Clarify when a catch block is required in conjunction with finally.
I've added entities for throw, try, catch and finally so that they can be more
easily used and linked to the right sections. This has resulted in a
considerable amount of code churn in exceptions.xml, as everything is now in a
simplesect or sect1 (and the sect1 had to be moved ahead of the simplesects,
which has no impact on formatting, but is required for the document to be
well-formed).

Fixes doc bug #69026 (conflict between documentation and behaviour on finally).


git-svn-id: https://svn.php.net/repository/phpdoc/en/trunk@335965 c90b9560-bf6c-de11-be94-00142212c4b1