What's in a name?
The Macintosh operating systems changed their name a few times since
version 10. We follow suite, and use the proper name whenever we refer
to a certain version, and use “macOS” as generic term, since “Mac OS X”
and “OS X” versions are already unsupported, or loose support at the end
of this year.

git-svn-id: https://svn.php.net/repository/phpdoc/en/trunk@345139 c90b9560-bf6c-de11-be94-00142212c4b1