Document potentially insecure IMAP mailbox names
We add a warning to all parameters which may be affected,
and also document the new ini option which prevents RCE.
We intentionally keep the wording backwards (“unless … is
disabled”) to also cater to older PHP version which do not
implement the ini option.

Cf. <https://bugs.php.net/bug.php?id=77153>.

git-svn-id: https://svn.php.net/repository/phpdoc/en/trunk@346027 c90b9560-bf6c-de11-be94-00142212c4b1