Fix Bug #79314 - Document CLI Server Workers
The PHP_CLI_SERVER_WORKERS feature, which enables the built-in web server to fork multiple workers
for handling concurrent requests, is not documented. I'm only including this in the changelog for
the commandline features page and not the migration guide as per previous discussions. The feature
was intended for running opcache tests in QA and shouldn't be advertised. However, documenting its
existence is still necessary.


git-svn-id: https://svn.php.net/repository/phpdoc/en/trunk@349267 c90b9560-bf6c-de11-be94-00142212c4b1